pyproject.toml
[project]
name = "battery-lab-deepagents-starter"
version = "0.1.0"
description = "Battery Lab Assistant starter built on deepagents and deep-agents-ui."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepagents>=0.2.6",
    "langchain-openai>=1.0.2",
    "langgraph-cli[inmem]>=0.1.55",
    "numpy>=2.2.1",
    "openai>=2.28.0",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.11.0",
]
pdf = [
    "reportlab>=4.4.10",
    "pdfplumber>=0.11.9",
    "pypdf>=6.9.1",
]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["battery_agent"]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
pythonpath = ["."]
